+++ lib/src/queries/iterable.dart
@@ -5,7 +5,7 @@ class _Iterable<E> extends IterableBase<E> {
 
   _Iterable(Iterator<E> iterator) {
     if (iterator == null) {
-      throw new ArgumentError.notNull("iterator");
+      throw ArgumentError.notNull("iterator");
     }
 
     _iterator = iterator;